SystemInfo fixed db model
fixed systeminfo class for the db model, added new class for system update. updated index and sysinfo pages to reflect different classes
