+++ arm/models/models.py
@@ -438,32 +438,21 @@ class UISettings(db.Model):
                 return_dict[str(key)] = str(value)
         return return_dict
 
-class SystemInfo():
+class SystemInfo(db.Model):
     """
-    Class to hold the system information
+    Class to hold the system (server) information
     """
+    id = db.Column(db.Integer, index=True, primary_key=True)
+    name = db.Column(db.String(100))
+    cpu = db.Column(db.String(20))
+    description = db.Column(db.Unicode(200))
+    mem_total = db.Column(db.Float())
 
-    cpu_name = "N/A"
-    cpu_util = 0.0
-    cpu_temp = 0.0
-    #memory = {}
-    memory_total = 0.0
-    memory_free = 0.0
-    memory_used = 0.0
-    memory_percent = 0.0
-    #hard drive space
-    storage_transcode_free = 0
-    storage_transcode_percent = 0.0
-    storage_completed_free = 0
-    storage_completed_percent = 0.0
-
-    def __init__(self):
+    def __init__(self, name="ARM Server", description="Automatic Ripping Machine main server"):
         self.get_cpu_info()
-        self.get_cpu_util()
-        self.get_cpu_temp()
         self.get_memory()
-        self.storage_transcode_free, self.storage_transcode_percent = self.get_disk_space(cfg['TRANSCODE_PATH'])
-        self.storage_completed_free, self.storage_completed_percent = self.get_disk_space(cfg['COMPLETED_PATH'])
+        self.name = name
+        self.description = description
 
     def get_cpu_info(self):
         """
@@ -471,9 +460,9 @@ class SystemInfo():
         ideally want to return {name} @ {speed} Ghz
         """
         if platform.system() == "Windows":
-            self.cpu_name = platform.processor()
+            self.cpu = platform.processor()
         elif platform.system() == "Darwin":
-            self.cpu_name = subprocess.check_output(['/usr/sbin/sysctl', "-n", "machdep.cpu.brand_string"]).strip()
+            self.cpu = subprocess.check_output(['/usr/sbin/sysctl', "-n", "machdep.cpu.brand_string"]).strip()
         elif platform.system() == "Linux":
             command = "cat /proc/cpuinfo"
             fulldump = str(subprocess.check_output(command, shell=True).strip())
@@ -485,8 +474,7 @@ class SystemInfo():
                 speeds = speeds.replace('\\n', ' ')
                 speeds = speeds.replace('\\t', ' ')
                 speeds = speeds.replace('model name :', '')
-                self.cpu_name = speeds
-
+                self.cpu = speeds
             # AMD CPU
             amd_name_full = re.search(r"model name\\t: (.*?)\\n", fulldump)
             if amd_name_full:
@@ -494,48 +482,16 @@ class SystemInfo():
                 amd_mhz = re.search(r"cpu MHz(?:\\t)*: ([.0-9]*)\\n", fulldump)  # noqa: W605
                 if amd_mhz:
                     amd_ghz = round(float(amd_mhz.group(1)) / 1000, 2)  # this is a good idea
-                    self.cpu_name = str(amd_name) + " @ " + str(amd_ghz) + " GHz"
-
-    def get_cpu_util(self):
-        try:
-            self.cpu_util = psutil.cpu_percent()
-        except EnvironmentError:
-            self.cpu_util = 0
-
-    def get_cpu_temp(self):
-        try:
-            temps = psutil.sensors_temperatures()
-            self.cpu_temp = temps['coretemp'][0][1]
-        except EnvironmentError:
-            self.cpu_temp = 0
+                    self.cpu = str(amd_name) + " @ " + str(amd_ghz) + " GHz"
+        else:
+            self.cpu = "N/A"
 
     def get_memory(self):
         try:
             memory = psutil.virtual_memory()
-            self.memory_total = round(memory.total / 1073741824, 1)
-            self.memory_free = round(memory.available / 1073741824, 1)
-            self.memory_used = round(memory.used / 1073741824, 1)
-            self.memory_percent = memory.percent
+            self.mem_total = round(memory.total / 1073741824, 1)
         except EnvironmentError:
-            self.memory_total = 0
-            self.memory_free = 0
-            self.memory_used = 0
-            self.memory_percent = 0
-
-    def get_disk_space(self, filepath):
-        # Hard drive space
-        try:
-            disk_space = psutil.disk_usage(filepath).free
-            disk_space = round(disk_space / 1073741824, 1)
-            disk_percent = psutil.disk_usage(filepath).percent
-        except FileNotFoundError:
-            disk_space = 0
-            disk_percent = 0
-            app.logger.debug("ARM folders not found")
-            flash("There was a problem accessing the ARM folders. Please make sure you have setup ARM<br/>"
-                  "Setup can be started by visiting <a href=\"/setup\">setup page</a> ARM will not work correctly until"
-                  "you have added an admin account", "danger")
-        return disk_space, disk_percent
+            self.mem_total = 0
 
 class SystemDrives(db.Model):
     """
