+++ arm/ui/templates/index.html
@@ -68,11 +68,11 @@
             </div>
             <ul class="list-group list-group-flush">
               <li class="list-group-item">
-                  CPU: {{ server.cpu_name  }}
+                  CPU: {{ server.cpu  }}
               </li>
               <li class="list-group-item">
                   {% if cpu_temp is not none %}
-                      <section class="{{ ' bg-danger text-white' if server.cpu_temp > 70 }}">CPU Temp: {{ server.cpu_temp }}&#176;C</section>
+                      <section class="{{ ' bg-danger text-white' if serverutil.cpu_temp > 70 }}">CPU Temp: {{ serverutil.cpu_temp }}&#176;C</section>
                   {% else %}
                     <section> Temp: not reported </section>
                   {% endif %}
@@ -80,9 +80,9 @@
               <li class="list-group-item">
                   <h6 class="progress-label" style="float: left;margin-right: 1em;">Usage: </h6>
                   <div class="progress">
-                      <div class="progress-bar{{ ' bg-danger' if server.cpu_util > 80 }}" role="progressbar"
-                           style="width: {{ ' 10' if server.cpu_util < 10 else server.cpu_util }}%;"
-                           aria-valuenow="{{ server.cpu_util }}" aria-valuemin="10" aria-valuemax="100">{{ server.cpu_util }}%
+                      <div class="progress-bar{{ ' bg-danger' if serverutil.cpu_util > 80 }}" role="progressbar"
+                           style="width: {{ ' 10' if serverutil.cpu_util < 10 else serverutil.cpu_util }}%;"
+                           aria-valuenow="{{ serverutil.cpu_util }}" aria-valuemin="10" aria-valuemax="100">{{ server.cpu_util }}%
                       </div>
                   </div>
               </li>
@@ -97,15 +97,15 @@
             <ul class="list-group list-group-flush">
                 <li class="list-group-item">
                   Total: {{ server.memory_total }} GB
-                  <br>Free:  {{ server.memory_free }} GB
-                  <br>Used:  {{ server.memory_used }} GB
+                  <br>Free:  {{ serverutil.memory_free }} GB
+                  <br>Used:  {{ serverutil.memory_used }} GB
                 </li>
                 <li class="list-group-item">
                     <h6 class="progress-label" style="float: left;margin-right: 1em;">Usage: </h6>
                     <div class="progress">
-                      <div class="progress-bar{{ ' bg-danger' if server.memory_percent > 80 }}" role="progressbar"
-                           style="width: {{ server.memory_percent }}%;" aria-valuenow="{{ server.memory_percent }}" aria-valuemin="0"
-                           aria-valuemax="100">{{ server.memory_percent }}%
+                      <div class="progress-bar{{ ' bg-danger' if serverutil.memory_percent > 80 }}" role="progressbar"
+                           style="width: {{ serverutil.memory_percent }}%;" aria-valuenow="{{ serverutil.memory_percent }}" aria-valuemin="0"
+                           aria-valuemax="100">{{ serverutil.memory_percent }}%
                       </div>
                     </div>
                   </li>
@@ -122,12 +122,12 @@
                     {% if server.storage_transcode_free != 0 %}
                     <h6 class="progress-label" style="float: left;margin-right: 1em;">Transcode: </h6>
                     <div class="progress">
-                        <div class="progress-bar{{ ' bg-danger' if server.storage_transcode_percent > 80 }}" role="progressbar"
-                             style="width: {{ server.storage_transcode_percent }}%;" aria-valuenow="{{ server.storage_transcode_percent }}"
-                             aria-valuemin="0" aria-valuemax="100">{{ server.storage_transcode_percent }}%
+                        <div class="progress-bar{{ ' bg-danger' if serverutil.storage_transcode_percent > 80 }}" role="progressbar"
+                             style="width: {{ serverutil.storage_transcode_percent }}%;" aria-valuenow="{{ serverutil.storage_transcode_percent }}"
+                             aria-valuemin="0" aria-valuemax="100">{{ serverutil.storage_transcode_percent }}%
                         </div>
                     </div>
-                    <br> Free Space: {{ server.storage_completed_free }} GB
+                    <br> Free Space: {{ serverutil.storage_completed_free }} GB
                     {% else %}
                     Transcode: Unable to get data on path
                     {% endif %}
@@ -137,9 +137,9 @@
                     {% if server.storage_completed_free != 0 %}
                     <h6 class="progress-label" style="float: left;margin-right: 1em;">Completed: </h6>
                     <div class="progress">
-                        <div class="progress-bar{{ ' bg-danger' if server.storage_completed_percent > 80 }}" role="progressbar"
-                             style="width: {{ server.storage_completed_percent }}%;" aria-valuenow="{{ server.storage_completed_percent }}"
-                             aria-valuemin="0" aria-valuemax="100">{{ server.storage_completed_percent }}%
+                        <div class="progress-bar{{ ' bg-danger' if serverutil.storage_completed_percent > 80 }}" role="progressbar"
+                             style="width: {{ serverutil.storage_completed_percent }}%;" aria-valuenow="{{ serverutil.storage_completed_percent }}"
+                             aria-valuemin="0" aria-valuemax="100">{{ serverutil.storage_completed_percent }}%
                         </div>
                     </div>
                     <br> Free Space: {{ server.storage_completed_free }} GB
